requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
loguru
httpx
python-socketio[asyncio_client]
